Tests:  - File request test  - Testing configuration fix (to junit 5)

+++ Back/MarketBackend/src/test/java/com/pashenko/marketbackend/filestotage/hashedstorage/AbstractHashedStorageTest.java
@@ -31,11 +31,13 @@ public class AbstractHashedStorageTest extends AbstractStorageTest {
     @BeforeAll
     static void setup() throws IOException {
         store = new HashedFileStorageService(hashCalc, root.toString());
-        Files.createFile(testFile.toAbsolutePath());
-        try (BufferedWriter bw = new BufferedWriter(new FileWriter(testFile.toAbsolutePath().toFile()))){
-            bw.write(testContent);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
+        if(!Files.exists(testFile)){
+            Files.createFile(testFile.toAbsolutePath());
+            try (BufferedWriter bw = new BufferedWriter(new FileWriter(testFile.toAbsolutePath().toFile()))){
+                bw.write(testContent);
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
         }
     }
